@@ -1,56 +1,29 @@
 {
-  "name": "vue2-laravel-custom-pagination",
-  "version": "1.0.5",
-  "description": "A Laravel Paginator for vue2 components",
+  "name": "vue3-laravel-custom-pagination",
+  "version": "2.0.0",
+  "description": "A Laravel Paginator for vue3 components",
   "author": "Bruno Gritti <glbrunogl@gmail.com>",
-
   "main": "dist/vue-laravel-pagination.ssr.js",
   "browser": "dist/vue-laravel-pagination.esm.js",
   "module": "dist/vue-laravel-pagination.esm.js",
   "unpkg": "dist/vue-laravel-pagination.min.js",
-
   "files": [
     "dist/*",
     "src/**/*.vue"
   ],
   "sideEffects": false,
-
   "scripts": {
     "serve": "vue-cli-service serve dev/serve.js",
     "prebuild": "rimraf ./dist",
     "build": "cross-env NODE_ENV=production rollup --config build/rollup.config.js"
   },
-
   "keywords": [
     "vue",
     "pagination",
     "laravel"
   ],
-
   "dependencies": {
-  },
-
-  "devDependencies": {
-    "@babel/core": "^7.14.6",
-    "@babel/preset-env": "^7.14.7",
-    "@rollup/plugin-alias": "^3.1.2",
-    "@rollup/plugin-babel": "^5.3.0",
-    "@rollup/plugin-commonjs": "^14.0.0",
-    "@rollup/plugin-node-resolve": "^9.0.0",
-    "@rollup/plugin-replace": "^2.4.2",
-    "@vue/cli-plugin-babel": "^4.5.13",
-    "@vue/cli-service": "^4.5.13",
-    "cross-env": "^7.0.3",
-    "minimist": "^1.2.5",
-    "rimraf": "^3.0.2",
-    "rollup": "^2.52.8",
-    "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-vue": "^5.1.9",
-    "vue": "^2.6.14",
-    "vue-template-compiler": "^2.6.14"
-  },
-  "peerDependencies": {
-    "vue": "^2.6.14"
+    "vue": "^3.4.23"
   },
   "engines": {
     "node": ">=12"
