@@ -1,42 +1,45 @@
 <template>
-<div class="btn-paginate">
-    <!-- Fazer v-if pra ocultar essa seção -->
-    <small> Mostrando {{from}} - {{to}} de {{total}} encontrados.</small>
-    <!-- Fazer classe dinamica com props -->
-    <button class="btn" :disabled="from == 1"
-        @click="prevPage">&laquo; Anterior
-    </button>
-    <button class="btn" :disabled="to == total"
-        @click="nextPage">Próximo &raquo;
-    </button>
+<div class="btn-paginate" v-if="total === 0">
+  <small> Mostrando {{from}} - {{to}} de {{total}} encontrados.</small>
+  <!-- Fazer classe dinamica com props -->
+  <button class="btn" :disabled="from === 1" @click="prevPage">&laquo; Anterior
+  </button>
+  <button class="btn" :disabled="to === total" @click="nextPage">Próximo &raquo;
+  </button>
 </div>
 </template>
 
 <script>
+
+import {ref} from "vue";
+
 export default {
-    props:['from', 'to', 'total', 'perPage'],
-    data() {
-      	return {
-            limit: this.perPage || 10,
-            page: 1
-      	};
-    },
+  props:['from', 'to', 'total'],
+  setup(props, {emit}) {
 
-    methods: {
-        prevPage: function(){
-            this.page--;
-            this.$emit('page-changed');
-        },
+    // let limit = ref(props.perPage ?? 10 );
+    let page = ref(1);
 
-        nextPage: function(){
-            this.page++;
-            this.$emit('page-changed');
-        },
-        
-        setPage: function(page){
-            this.page = page;
-        },
+    function prevPage(){
+      page.value--;
+      setPage(page.value);
+      emit('page-changed', {page: page.value});
     }
+
+    function nextPage(){
+      page.value++;
+      setPage(page.value);
+      emit('page-changed', {page: page.value});
+    }
+
+    function setPage(page){
+      page.value = page;
+    }
+
+    return {
+      prevPage
+    }
+  }
 }
 </script>
 
